[1/x] split project page into project info component
