+++ src/components/v2/V2Project/V2ProjectSettings.tsx
@@ -1,0 +1,5 @@
+const V2ProjectSettings = () => {
+  return <div>V2ProjectSettings</div>
+}
+
+export default V2ProjectSettings
