+++ src/constants/featureFlags.ts
@@ -3,4 +3,5 @@ export const FEATURE_FLAGS = {
   NFT_REWARDS: 'nftRewards',
   VENFT: 'veNft',
   VENFT_CREATOR: 'veNftCreator',
+  SETTINGS_PAGE: 'settingsPage',
 }
