+++ src/models/project-visibility.ts
@@ -1,2 +1,3 @@
 export type ProjectState = 'active' | 'archived'
 export type ProjectCategory = 'all' | 'holdings' | 'myprojects' | 'trending' //| 'metaverse' | 'defi' | 'nft' | 'web3'
+export type ProjectPage = 'info' | 'settings'
